move parallel interleave checkpoint tests to kernel tests
